requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver
alembic==1.12.1  # Database migrations

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing API endpoints

# Optional: for production deployment
gunicorn==21.2.0

# Optional: for caching
redis==5.0.1

# Optional: for file uploads and image processing
pillow==10.1.0
boto3==1.34.0  # AWS S3 for file storage 